# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Rest API for Payment Processing.",
        "title": "Payment Processor",
        "termsOfService": "https://github.com/sesaquecruz/go-payment-processor",
        "contact": {
            "name": "Support",
            "url": "https://github.com/sesaquecruz/go-payment-processor"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/sesaquecruz/go-payment-processor"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/payments/process": {
            "post": {
                "security": [
                    {
                        "Bearer token": []
                    }
                ],
                "description": "Process a payment transaction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Process a payment",
                "parameters": [
                    {
                        "description": "Transaction",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Transaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Payment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.HttpError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.HttpError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.HttpError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.Payment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.Transaction": {
            "type": "object",
            "required": [
                "acquirer_name",
                "card_token",
                "purchase_installments",
                "purchase_items",
                "purchase_value",
                "store_address",
                "store_cep",
                "store_identification"
            ],
            "properties": {
                "acquirer_name": {
                    "type": "string"
                },
                "card_token": {
                    "type": "string"
                },
                "purchase_installments": {
                    "type": "integer"
                },
                "purchase_items": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "purchase_value": {
                    "type": "number"
                },
                "store_address": {
                    "type": "string"
                },
                "store_cep": {
                    "type": "string"
                },
                "store_identification": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer token": {
            "description": "Authorization Token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}